Move goto_statet definition to its own file
Having one class per header file makes it easier to navigate the code
and find what we want.
